feat(core): convert date in values to unixtime
