+++ packages/core/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@react-native-karte/core",
-  "version": "1.2.0",
+  "version": "1.3.0",
   "description": "KARTE SDK Core for React Native.",
   "main": "lib/commonjs/index",
   "module": "lib/module/index",
